deployment/requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
astunparse==1.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
filelock==3.8.2
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
libclang==15.0.6.1
lingua-language-detector==1.1.3
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
pytrends==4.9.0
pytz==2022.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
snscrape==0.5.0.20230113
soupsieve==2.3.2.post1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.13
vaderSentiment==3.3.2
Werkzeug==2.2.3
wrapt==1.14.1
-e .
